Button positions now check for - number or + number, added info to main
Buttons now check for position based on negative or positive numbers, not a specific number ( was 780 ). My controller was outputting -1000 and 1000 on the left and right switch / buttons, so the new change should allow it to work with all DJI Phantom 2 controllers. Also added a ( Digital Stick Center ) for people who wish to play with all springs removed ( DJI controller springs are very strong ). Moving the camera slider all the way to max ( right ) and flipping the right and left switch to the bottom positions will trick the Vjoy controller into thinking both sticks are positioned at center. Later I will add this as an option that users can turn on or off. Updated the ( mDjiController.zip ) to now have the correct compiled program in it ( was a very old version that output the sticks to 0 - 2000 and was not multiplying it by 16 ).