Added some new features.
* Automatic Port Scanning.
* File creation to save and load up settings data.
* Cleaned up information presentation and simplified the users process.
*Added a ClearScreen to create a simple UI and make Logging give the appearance of live updates.

I intend to make the port scan process fully automated for a plug and play experience. 

I also intent to test this on other DJI controllers once it is finish and maybe eventually turn it into a full interface for using all USB available DJI controllers on Simulators.

Work in progress...