Added Auto Controller Discovery
Using (Enumser) to scan the COM ports, then sift through for key word "DJI" to list only DJI devices and auto connect. 

Cleaned up the process of starting up a bit more, added a feature to choose if you create a save file for auto start after your initial setup of controller.

Made changes to the "Readme" to bring them up-to-date with the changes and to make it better for SEO to help the software get found.

Added a more modern photo (in_use.png) with a modern game to show that this works with games like "LiftOff".

More changes to the UI and instructions as the user sets up the program for the first time.